requirements.txt
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
itsdangerous==0.24
Jinja2==2.9.6
kafka-python==1.3.5
MarkupSafe==1.0
six==1.11.0
Werkzeug==0.12.2
